Move most of WorkerHarness into HostRuntime.
